arreglo a test api persona
se cambio el modo en el que se borra, ya que si queda expuesto
en la api se podria borrar y dejar resultados indeseados.
por lo que ahora se borra directamente de la base de datos sin
pasar por la API
